// src/clients/models/planet.ts
/* eslint-disable @typescript-eslint/naming-convention */
export type Planet = {
	name: string;
	diameter: string;
	rotation_period: string;
	orbital_period: string;
	gravity: string;
	population: string | null;
	climate: string;
	terrain: string;
	surface_water: string;
	residents: string[];
	films: string[];
	url: string;
	created: string;
	edited: string;
};